Split Wonder into WonderType and WonderSide
It was always a bit nasty that we had enum values like
"ColossusOfRhodesA". We also need to be able to randomly choose wonders
for players when initialising a game, and need to prevent one player
getting ColossusOfRhodesA while another gets ColossusOfRhodesB.

Split into a type and a side, and introduce a struct, "WonderBoard" that
holds the two.