Introduce VisibleGame for passing public game state to various functions
In preparation for supporting borrowing resources from neighbours, we're
going to need access to the "visible" game state (ie. the state all
players can see, ie. excluding players' hands etc). We're also going to
need access to players immediately to the left and right of the current
player. Introduce VisibleGame is hold all this information.