PlayingAlgorithm.get_next_action() needs to mutate self
When we start on slightly more sophisticated playing algorithms, they'll
need to maintain state and mutate it when calculating their next
action (eg. keeping track of which cards are probably where).

The algorithms still need access to Player so they can see the current
hand and call can_play(), options_for_card() etc. This is a problem
because we need to pass an immutable borrow of Player while calling
PlayingAlgorithm.get_next_action() with a mutable borrow of self. With
PlayingAlgorithm inside Player this leads to a mutable borrow at the
same time as an immutable borrow of Player. Therefore, we need to
separate the instance of PlayingAlgorithm from Player and store it
alongside instead (either that or clone all the player data each call,
which is expensive).