Rename Resources to Cost and introduce Resource
We were using the Resources struct to represent both the cost of a card
and the "power" of brown, grey, and yellow cards (eg. produces 1 wood).
While they're obviously very related, they're actually quite different,
and we had accumulated quite a lot of awkward code to try to work with
this situation (eg. see methods in resources.rs such as max and split).

Rename Resources to Cost and continue using it to represent the cost of
cards. Introduce a simple Resource enum that represents each of the
basic resources, and use that for "power" and borrowing calculations.