Record the resource, not the index, in a Borrow
The Borrow struct stores the index of the resource on the card that is
being borrowed. It's awkward and not actually very useful. Instead,
store the resource. For cards with two instances of the same resource,
we're unable to distinguish which instance is being borrowed, but we
don't need that information.

This is also a step towards being able to check a build action is valid
without having to run the Player::options_for_card function again, which
is inefficient and prevents certain optimisations of options_for_card.